configured mongo db and removed h2 configuration
