--- src/main/kotlin/de/tom/demo/springrestfulwebservice/Application.kt
@@ -1,13 +1,0 @@
-package de.tom.demo.springrestfulwebservice
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class Application
-
-fun main(args: Array<String>) {
-	runApplication<Application>(*args)
-}
-
-
