--- src/main/kotlin/de/tom/demo/springrestfulwebservice/config/PasswordEncoder.kt
@@ -1,11 +1,0 @@
-package de.tom.demo.springrestfulwebservice.config
-
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder
-
-@Configuration
-class PasswordEncoder {
-    @Bean
-    fun bCryptPasswordEncoder(): BCryptPasswordEncoder = BCryptPasswordEncoder()
-}
