+++ src/main/kotlin/de/tom/demo/taskapp/Application.kt
@@ -1,0 +1,13 @@
+package de.tom.demo.taskapp
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class Application
+
+fun main(args: Array<String>) {
+	runApplication<Application>(*args)
+}
+
+
