+++ src/main/kotlin/de/tom/demo/taskapp/config/PasswordEncoder.kt
@@ -1,0 +1,11 @@
+package de.tom.demo.taskapp.config
+
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder
+
+@Configuration
+class PasswordEncoder {
+    @Bean
+    fun bCryptPasswordEncoder(): BCryptPasswordEncoder = BCryptPasswordEncoder()
+}
