+++ src/main/kotlin/de/tom/demo/taskapp/entities/tasks/TaskRepository.kt
@@ -1,0 +1,12 @@
+package de.tom.demo.taskapp.entities.tasks
+
+import de.tom.demo.taskapp.entities.Task
+import org.springframework.data.mongodb.repository.MongoRepository
+import org.springframework.stereotype.Repository
+
+/**
+ *  JPA Repository: to interact with H2 database
+ */
+@Repository
+interface TaskRepository : MongoRepository<Task, String> {
+}
