--- src/main/resources/application-dev-h2.properties
@@ -1,12 +1,0 @@
-spring.datasource.driver-class-name=org.h2.Driver
-spring.datasource.url=jdbc:h2:file:./data/testdb
-spring.datasource.username=sa
-spring.datasource.password=password
-spring.datasource.schema=classpath:sql/schema.sql
-spring.datasource.initialization-mode=always
-
-# Add error message to the server response
-server.error.include-message=always
-server.port=5000
-
-#logging.level.root=DEBUG
