+++ src/main/resources/application-dev.properties
@@ -1,0 +1,8 @@
+spring.data.mongodb.uri=mongodb+srv://tomwey:lareunion@tomweycluster.ehxug.mongodb.net/taskapp
+spring.data.mongodb.database=taskapp
+
+# Add error message to the server response
+server.error.include-message=always
+server.port=5000
+
+#logging.level.root=DEBUG
