+++ src/main/resources/application.properties
@@ -1,1 +1,1 @@
-spring.profiles.active=dev-h2
+spring.profiles.active=dev
