--- src/test/kotlin/de/tom/demo/springrestfulwebservice/ApplicationTests.kt
@@ -1,15 +1,0 @@
-package de.tom.demo.springrestfulwebservice
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class ApplicationTests {
-	// a simple sanity check test that will fail if the application context cannot start.
-
-	@Test
-	fun contextLoads() {
-		println("ApplicationTests")
-	}
-
-}
